auth/requirements.txt
django-environ===0.12.0
django===5.2.8
djangorestframework-simplejwt===5.5.1
djangorestframework===3.16.1
uvicorn===0.38.0
django-cors-headers===4.9.0
psycopg[binary]===3.2.12
django-redis===6.0.0
whitenoise===6.11.0
bcrypt===5.0.0
cryptography===46.0.3
django-health-check